Revert treating end of file as a ending a python func
This caused issues with the the first plain python function starting
with def not ending until the end of the file, causing the whole file
to be treated as python.

For now we revert to old behavior, and python funcs that end in end of
file are not properly highlighted.

We need a better solution...